Add step to generate files using DAG generator to CI

+++ include/multi_file/generator.py
@@ -1,7 +1,7 @@
 import os, fileinput, shutil, json
 
-DAGS_FOLDER = "/usr/local/airflow/dags"
-module_path = f"{DAGS_FOLDER}/examples/generator/multi_file"
+DAGS_FOLDER = "./dags"
+module_path = f"./include/multi_file"
 config_filepath = f"{module_path}/configs/"
 generated_dags_path = f"{DAGS_FOLDER}/examples/generated_dags"
 
@@ -15,14 +15,14 @@ for filename in os.listdir(config_filepath):
     shutil.copyfile(f"{module_path}/template.py", new_filename)
 
     for line in fileinput.input(new_filename, inplace=True):
-        line = line.replace("dag_id_to_replace", "'" + config["dag_id"] + "'")
-        line = line.replace("schedule_to_replace", config["schedule"])
-        line = line.replace("bash_command_to_replace", config["bash_command"])
-        line = line.replace("env_var_to_replace", config["env_var"])
+        line = line.replace("$dag_id", config["dag_id"])
+        line = line.replace("$schedule", config["schedule"])
+        line = line.replace("$bash_command", config["bash_command"])
+        line = line.replace("$env_var", {"123": "123"})
         print(line, end="")
 
-    generated_dags.append(f"{config["dag_id"]}.py")
+        generated_dags.append(f"{config["dag_id"]}.py")
 
 for filename in os.listdir(generated_dags_path):
-    if filename not in generated_dags:
+    if filename.endswith(".py") and filename not in generated_dags:
         os.remove(f"{generated_dags_path}/{filename}")
